+++ services/policy/extract_policy_by_cloudTrail.py
@@ -1,9 +1,9 @@
 import os
 from dotenv import load_dotenv
-from utils.policy.common_utils import load_json, merge_policies, map_etc
-from utils.policy.s3_policy_mapper import s3_policy_mapper
-from utils.policy.ec2_policy_mapper import ec2_policy_mapper
-from utils.policy.iam_policy_mapper import iam_policy_mapper
+from services.policy.common_utils import load_json, merge_policies, map_etc
+from services.policy.s3_policy_mapper import s3_policy_mapper
+from services.policy.ec2_policy_mapper import ec2_policy_mapper
+from services.policy.iam_policy_mapper import iam_policy_mapper
 
 load_dotenv()
 
@@ -59,7 +59,7 @@ def making_policy(log_entry):
 def extract_policy_by_cloudTrail():
     # ES에서 가져와야 하는 로그. 일단 sample data로 대체
     iam_policy_dir = os.getenv("IAM_POLICY_DIR_PATH")
-    file_path = os.path.join(iam_policy_dir, "/src/sample_data/event_history.json")
+    file_path = os.path.join(iam_policy_dir, "src/sample_data/event_history.json")
 
     logs = load_json(file_path).get("Records", [])
     if not isinstance(logs, list):
