+++ services/policy/s3_policy_mapper.py
@@ -1,4 +1,4 @@
-from utils.policy.common_utils import generate_least_privilege_policy
+from services.policy.common_utils import generate_least_privilege_policy
 
 def s3_policy_mapper(log, policy_data):
     """S3 정책 생성."""
