+++ services/user_service.py
@@ -1,6 +1,7 @@
 from fastapi import Depends
 from repositories.user_repository import UserRepository
 
+
 class UserService:
     def __init__(self, user_repository: UserRepository = Depends()):
         self.user_repository = user_repository
